download_metadata {geobr}	R Documentation

Support function to download metadata internally used in geobr

Description

Support function to download metadata internally used in geobr

Usage

download_metadata()

Examples

## Not run:  if (interactive()) {
df <- download_metadata()
}
## End(Not run)

[Package geobr version 1.9.1 Index]