download_gpkg {geobr}	R Documentation

Download geopackage to tempdir

Description

Download geopackage to tempdir

Usage

download_gpkg(
  file_url = parent.frame()$file_url,
  showProgress = parent.frame()$showProgress,
  cache = parent.frame()$cache
)

Arguments

file_url
A string with the file_url address of a geobr dataset

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

[Package geobr version 1.9.1 Index]