Support function to download metadata internally used in geobr

Support function to download metadata internally used in geobr

Usage

download_metadata()

Examples

if (FALSE)  if (interactive()) {
df <- download_metadata()
} # \dontrun{}