geobr {geobr}	R Documentation

geobr: Download Official Spatial Data Sets of Brazil

Description

Easy access to official spatial data sets of Brazil as 'sf' objects in R. The package includes a wide range of geospatial data available at various geographic scales and for various years with harmonized attributes, projection and fixed topology.

Usage

Please check the vignettes for more on the package usage:

Introduction to geobr (R) on the website.

Author(s)

Maintainer: Rafael H. M. Pereira rafa.pereira.br@gmail.com (ORCID)

Authors:

Caio Nogueira Goncalves

Other contributors:

Paulo Henrique Fernandes de Araujo [contributor]

Guilherme Duarte Carvalho [contributor]

Rodrigo Almeida de Arruda [contributor]

Igor Nascimento [contributor]

Barbara Santiago Pedreira da Costa [contributor]

Welligtton Silva Cavedo [contributor]

Pedro R. Andrade [contributor]

Alan da Silva [contributor]

Carlos Kauê Vieira Braga [contributor]

Carl Schmertmann [contributor]

Alessandro Samuel-Rosa [contributor]

Daniel Ferreira [contributor]

Marcus Saraiva [contributor]

Beatriz Milz (ORCID) [contributor]

Ipea - Institue for Applied Economic Research [copyright holder, funder]

See Also

Useful links:

https://ipeagit.github.io/geobr/

https://github.com/ipeaGIT/geobr

Report bugs at https://github.com/ipeaGIT/geobr/issues

[Package geobr version 1.9.1 Index]