list_geobr {geobr}	R Documentation

List all data sets available in the geobr package

Description

Returns a data frame with all datasets available in the geobr package

Usage

list_geobr()

Value

A data.frame

See Also

Other support functions: lookup_muni()

Examples

df <- list_geobr()

[Package geobr version 1.9.1 Index]