Look up municipality codes and names

Input a municipality name or code and get the names and codes of the municipality's corresponding state, meso, micro, intermediate, and immediate regions

Usage

lookup_muni(name_muni = NULL, code_muni = NULL)

Arguments

name_muni

The municipality name to be looked up.

code_muni

The municipality code to be looked up.

Value

A data.frame with 13 columns identifying the geographies information of that municipality.

A data.frame

Details

Only available from 2010 Census data so far

See also

Examples

# Get lookup table for municipality Rio de Janeiro
mun <- lookup_muni(name_muni = "Rio de Janeiro")
#> Using year/date 2010
#> Returning results for municipality Rio de Janeiro

# Or you can get a lookup table for the same municipality searching for its code
mun <- lookup_muni(code_muni = 3304557)
#> Returning results for municipality Rio de Janeiro

# Get lookup table for all municipalities
mun_all <- lookup_muni(name_muni = "all")

# Or:
mun_all <- lookup_muni(code_muni = "all")