lookup_muni {geobr}	R Documentation

Look up municipality codes and names

Description

Input a municipality name or code and get the names and codes of the municipality's corresponding state, meso, micro, intermediate, and immediate regions

Usage

lookup_muni(name_muni = NULL, code_muni = NULL)

Arguments

name_muni
The municipality name to be looked up.

code_muni
The municipality code to be looked up.

Details

Only available from 2010 Census data so far

Value

A data.frame with 13 columns identifying the geographies information of that municipality.

A data.frame

See Also

Other support functions: list_geobr()

Examples

# Get lookup table for municipality Rio de Janeiro
mun <- lookup_muni(name_muni = "Rio de Janeiro")

# Or you can get a lookup table for the same municipality searching for its code
mun <- lookup_muni(code_muni = 3304557)

# Get lookup table for all municipalities
mun_all <- lookup_muni(name_muni = "all")

# Or:
mun_all <- lookup_muni(code_muni = "all")

[Package geobr version 1.9.1 Index]