read_capitals {geobr}	R Documentation

Download data of state capitals

Description

This function downloads either a spatial sf object with the location of the municipal seats (sede dos municipios) of state capitals, or a data.frame with the names and codes of state capitals. Data downloaded for the latest available year.

Usage

read_capitals(as_sf = TRUE, showProgress = TRUE)

Arguments

as_sf
Logic FALSE or TRUE, indicating whether the function should return a spatial data in sf format (Defaults to TRUE) or in a data.frame format without spatial information (FALSE).

showProgress
Logical. Defaults to TRUE display progress bar.

Value

An "sf" "data.frame" object or a "data.frame"

See Also

Other area functions: read_amazon(), read_biomes(), read_comparable_areas(), read_country(), read_disaster_risk_area(), read_health_facilities(), read_health_region(), read_immediate_region(), read_indigenous_land(), read_intermediate_region(), read_meso_region(), read_metro_area(), read_micro_region(), read_municipal_seat(), read_municipality(), read_neighborhood(), read_pop_arrangements(), read_region(), read_schools(), read_semiarid(), read_state(), read_statistical_grid(), read_urban_area(), read_urban_concentrations(), read_weighting_area()

Examples

# Read spatial data with the  municipal seats of state capitals
capitals_sf <- read_capitals(as_sf = TRUE)

# Read simple data.frame of state capitals
capitals_df <- read_capitals(as_sf = FALSE)

[Package geobr version 1.9.1 Index]