read_conservation_units {geobr}	R Documentation

Download spatial data of Brazilian environmental conservation units

Description

This data set covers the whole of Brazil and it includes the polygons of all conservation units present in Brazilian territory. The last update of the data was 09-2019. The original data comes from MMA and can be found at "http://mapas.mma.gov.br/i3geo/datadownload.htm".

Usage

read_conservation_units(
  date = 201909,
  simplified = TRUE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

date
Numeric. Date of the data in YYYYMM format. Defaults to 201909.

simplified
Logic FALSE or TRUE, indicating whether the function should return the data set with 'original' spatial resolution or a data set with 'simplified' geometry. Defaults to TRUE. For spatial analysis and statistics users should set simplified = FALSE. Borders have been simplified by removing vertices of borders using st_simplify{sf} preserving topology with a dTolerance of 100.

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See Also

Other general area functions: read_census_tract()

Examples

# Read conservation_units
b <- read_conservation_units(date = 201909)

[Package geobr version 1.9.1 Index]