Download geolocated data of health facilities

Data comes from the National Registry of Healthcare facilities (Cadastro Nacional de Estabelecimentos de Saude - CNES), originally collected by the Brazilian Ministry of Health. According to the Ministry of Health: "The coordinates of each facility were obtained by CNES and validated by means of space operations. These operations verify if the point is in the municipality, considering a radius of 5,000 meters. When the coordinate is not correct, further searches are done in other systems of the Ministry of Health and in web services like Google Maps. Finally, if the coordinates have been correctly obtained in this process, the coordinates of the municipal head office are used. The geocode source used is registered in the database in a specific column data_source. Periodically the coordinates are revised with the objective of improving the quality of the data." The date of the last data update is registered in the database in the columns date_update and year_update. More information in the CNES data set available at https://dados.gov.br/. These data use Geodetic reference system "SIRGAS2000" and CRS(4674).

Usage

read_health_facilities(date = NULL, showProgress = TRUE, cache = TRUE)

Arguments

date

Numeric. Date of the data in YYYYMM format. It defaults to NULL and reads the data from the latest date available.

showProgress

Logical. Defaults to TRUE display progress bar.

cache

Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See also

Examples

# Read all health facilities of the whole country
h <- read_health_facilities( date = 202303)
#> Using year/date 202303