read_health_region {geobr}	R Documentation

Download spatial data of Brazilian health regions and health macro regions

Description

Health regions are used to guide the the regional and state planning of health services. Macro health regions, in particular, are used to guide the planning of high complexity health services. These services involve larger economics of scale and are concentrated in few municipalities because they are generally more technology intensive, costly and face shortages of specialized professionals. A macro region comprises one or more health regions.

Usage

read_health_region(
  year = 2013,
  macro = FALSE,
  simplified = TRUE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year
Numeric. Year of the data in YYYY format. Defaults to 2013, the latest available.

macro
Logic. If FALSE (default), the function downloads health regions data. If TRUE, the function downloads macro regions data.

simplified
Logic FALSE or TRUE, indicating whether the function should return the data set with 'original' spatial resolution or a data set with 'simplified' geometry. Defaults to TRUE. For spatial analysis and statistics users should set simplified = FALSE. Borders have been simplified by removing vertices of borders using st_simplify{sf} preserving topology with a dTolerance of 100.

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See Also

Other area functions: read_amazon(), read_biomes(), read_capitals(), read_comparable_areas(), read_country(), read_disaster_risk_area(), read_health_facilities(), read_immediate_region(), read_indigenous_land(), read_intermediate_region(), read_meso_region(), read_metro_area(), read_micro_region(), read_municipal_seat(), read_municipality(), read_neighborhood(), read_pop_arrangements(), read_region(), read_schools(), read_semiarid(), read_state(), read_statistical_grid(), read_urban_area(), read_urban_concentrations(), read_weighting_area()

Examples

# Read all health regions for a given year
hr <- read_health_region( year=2013 )

# Read all macro health regions
mhr <- read_health_region( year=2013, macro =TRUE)

[Package geobr version 1.9.1 Index]