read_indigenous_land {geobr}	R Documentation

Download spatial data of indigenous lands in Brazil

Description

The data set covers the whole of Brazil and it includes indigenous lands from all ethnicities and in different stages of demarcation. The original data comes from the National Indian Foundation (FUNAI) and can be found at https://www.gov.br/funai/pt-br/atuacao/terras-indigenas/geoprocessamento-e-mapas. Although original data is updated monthly, the geobr package will only keep the data for a few months per year.

Usage

read_indigenous_land(
  date = 201907,
  simplified = TRUE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

date
Numeric. Date of the data in YYYYMM format. Defaults to 201907.

simplified
Logic FALSE or TRUE, indicating whether the function should return the data set with 'original' spatial resolution or a data set with 'simplified' geometry. Defaults to TRUE. For spatial analysis and statistics users should set simplified = FALSE. Borders have been simplified by removing vertices of borders using st_simplify{sf} preserving topology with a dTolerance of 100.

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See Also

Other area functions: read_amazon(), read_biomes(), read_capitals(), read_comparable_areas(), read_country(), read_disaster_risk_area(), read_health_facilities(), read_health_region(), read_immediate_region(), read_intermediate_region(), read_meso_region(), read_metro_area(), read_micro_region(), read_municipal_seat(), read_municipality(), read_neighborhood(), read_pop_arrangements(), read_region(), read_schools(), read_semiarid(), read_state(), read_statistical_grid(), read_urban_area(), read_urban_concentrations(), read_weighting_area()

Examples

# Read all indigenous land in an specific date
i <- read_indigenous_land(date=201907)

[Package geobr version 1.9.1 Index]