Download official data of municipal seats (sede dos municipios) in Brazil as an sf object.

This function reads the official data on the municipal seats (sede dos municipios) of Brazil. The data brings the spatial coordinates (lat lon) of of municipal seats for various years between 1872 and 2010. Orignal data were generated by Brazilian Institute of Geography and Statistics (IBGE).

read_municipal_seat(year = 2010, showProgress = TRUE)

Arguments

year
A year number in YYYY format (Defaults to 2010)

showProgress
Logical. Defaults to (TRUE) display progress bar

Examples

# \donttest{

library(geobr)

# Read municipal seats in an specific year
  m <- read_municipal_seat(year=1991)

#> Using year 1991

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Downloading: 770 B     Downloading: 770 B     Downloading: 1.8 kB     Downloading: 1.8 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 2 kB     Downloading: 2 kB     Downloading: 2.1 kB     Downloading: 2.1 kB     Downloading: 3.7 kB     Downloading: 3.7 kB     Downloading: 3.7 kB     Downloading: 3.7 kB     Downloading: 3.7 kB     Downloading: 3.7 kB     Downloading: 3.7 kB     Downloading: 3.7 kB     Downloading: 12 kB     Downloading: 12 kB     Downloading: 15 kB     Downloading: 15 kB     Downloading: 15 kB     Downloading: 15 kB     Downloading: 15 kB     Downloading: 15 kB     Downloading: 23 kB     Downloading: 23 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 31 kB     Downloading: 39 kB     Downloading: 39 kB     Downloading: 39 kB     Downloading: 39 kB     Downloading: 39 kB     Downloading: 39 kB     Downloading: 39 kB     Downloading: 39 kB     Downloading: 47 kB     Downloading: 47 kB     Downloading: 47 kB     Downloading: 47 kB     Downloading: 47 kB     Downloading: 47 kB     Downloading: 55 kB     Downloading: 55 kB     Downloading: 55 kB     Downloading: 55 kB     Downloading: 72 kB     Downloading: 72 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 80 kB     Downloading: 88 kB     Downloading: 88 kB     Downloading: 88 kB     Downloading: 88 kB     Downloading: 96 kB     Downloading: 96 kB     Downloading: 96 kB     Downloading: 96 kB     Downloading: 96 kB     Downloading: 96 kB     Downloading: 100 kB     Downloading: 100 kB     Downloading: 100 kB     Downloading: 100 kB     Downloading: 100 kB     Downloading: 100 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 120 kB     Downloading: 120 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 170 kB     Downloading: 170 kB     Downloading: 170 kB     Downloading: 170 kB     Downloading: 170 kB     Downloading: 170 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 190 kB     Downloading: 190 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 220 kB     Downloading: 220 kB     Downloading: 220 kB     Downloading: 220 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 270 kB     Downloading: 270 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     


# }