read_municipal_seat {geobr}	R Documentation

Download spatial data of municipal seats (sede dos municipios) in Brazil

Description

This function reads the official data on the municipal seats (sede dos municipios) of Brazil. The data brings the geographical coordinates (lat lon) of municipal seats for various years between 1872 and 2010. Original data were generated by Brazilian Institute of Geography and Statistics (IBGE).

Usage

read_municipal_seat(year = 2010, showProgress = TRUE, cache = TRUE)

Arguments

year
Numeric. Year of the data in YYYY format. Defaults to 2010.

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See Also

Other area functions: read_amazon(), read_biomes(), read_capitals(), read_comparable_areas(), read_country(), read_disaster_risk_area(), read_health_facilities(), read_health_region(), read_immediate_region(), read_indigenous_land(), read_intermediate_region(), read_meso_region(), read_metro_area(), read_micro_region(), read_municipality(), read_neighborhood(), read_pop_arrangements(), read_region(), read_schools(), read_semiarid(), read_state(), read_statistical_grid(), read_urban_area(), read_urban_concentrations(), read_weighting_area()

Examples

# Read municipal seats in an specific year
m <- read_municipal_seat(year = 1991)

[Package geobr version 1.9.1 Index]