Download spatial data of Brazil Regions

Data at scale 1:250,000, using Geodetic reference system "SIRGAS2000" and CRS(4674)

read_region(year = 2010, simplified = TRUE, showProgress = TRUE)

Arguments

year

Year of the data (defaults to 2010)

simplified

Logic FALSE or TRUE, indicating whether the function returns the data set with original' resolution or a data set with 'simplified' borders. Defaults to TRUE. For spatial analysis and statistics users should set simplified = FALSE. Borders have been simplified by removing vertices of borders using sf::st_simplify() preserving topology with a dTolerance of 100.

showProgress

Logical. Defaults to TRUE display progress bar

Value

An "sf" "data.frame" object

See also

Examples

if (FALSE)  if (interactive()) {
# Read specific year
reg <- read_region(year=2018)
}