read_schools {geobr}	R Documentation

Download geolocated data of schools

Description

Data comes from the School Census collected by INEP, the National Institute for Educational Studies and Research Anisio Teixeira. The date of the last data update is registered in the database in the column 'date_update'. These data uses Geodetic reference system "SIRGAS2000" and CRS(4674). The coordinates of each school if collected by INEP. Periodically the coordinates are revised with the objective of improving the quality of the data. More information available at https://www.gov.br/inep/pt-br/acesso-a-informacao/dados-abertos/inep-data/catalogo-de-escolas/

Usage

read_schools(year = 2020, showProgress = TRUE, cache = TRUE)

Arguments

year
Numeric. Year of the data in YYYY format. Defaults to 2020.

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See Also

Other area functions: read_amazon(), read_biomes(), read_capitals(), read_comparable_areas(), read_country(), read_disaster_risk_area(), read_health_facilities(), read_health_region(), read_immediate_region(), read_indigenous_land(), read_intermediate_region(), read_meso_region(), read_metro_area(), read_micro_region(), read_municipal_seat(), read_municipality(), read_neighborhood(), read_pop_arrangements(), read_region(), read_semiarid(), read_state(), read_statistical_grid(), read_urban_area(), read_urban_concentrations(), read_weighting_area()

Examples

# Read all schools in the country
s <- read_schools( year = 2020)

[Package geobr version 1.9.1 Index]