Download spatial data of Brazilian states

Data at scale 1:250,000, using Geodetic reference system "SIRGAS2000" and CRS(4674)

Usage

read_state(
  year = 2010,
  code_state = "all",
  simplified = TRUE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of the data in YYYY format. It defaults to NULL and reads the data from the latest year available.

code_state

The two-digit code of a state or a two-letter uppercase abbreviation (e.g. 33 or "RJ"). If code_state="all" (the default), the function downloads all states.

simplified

Logic FALSE or TRUE, indicating whether the function should return the data set with 'original' spatial resolution or a data set with 'simplified' geometry. Defaults to TRUE. For spatial analysis and statistics users should set simplified = FALSE. Borders have been simplified by removing vertices of borders using st_simplify{sf} preserving topology with a dTolerance of 100.

showProgress

Logical. Defaults to TRUE display progress bar.

cache

Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See also

Examples

# Read specific state at a given year
  uf <- read_state(code_state=12, year=2017)
#> Using year/date 2017

# Read specific state at a given year
  uf <- read_state(code_state="SC", year=2000)
#> Using year/date 2000

# Read all states at a given year
  ufs <- read_state(code_state="all", year=2010)
#> Using year/date 2010