read_statistical_grid {geobr}	R Documentation

Download spatial data of IBGE's statistical grid

Description

Data at scale 1:250,000, using Geodetic reference system "SIRGAS2000" and CRS(4674)

Usage

read_statistical_grid(
  code_grid,
  year = 2010,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

code_grid
If two-letter abbreviation or two-digit code of a state is passed, the function will load all grid quadrants that intersect with that state. If code_grid="all", the grid of the whole country will be loaded. Users may also pass a grid quadrant id to load an specific quadrant. Quadrant ids can be consulted at geobr::grid_state_correspondence_table.

year
Numeric. Year of the data in YYYY format. Defaults to 2010. The only year available thus far is 2010.

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See Also

Other area functions: read_amazon(), read_biomes(), read_capitals(), read_comparable_areas(), read_country(), read_disaster_risk_area(), read_health_facilities(), read_health_region(), read_immediate_region(), read_indigenous_land(), read_intermediate_region(), read_meso_region(), read_metro_area(), read_micro_region(), read_municipal_seat(), read_municipality(), read_neighborhood(), read_pop_arrangements(), read_region(), read_schools(), read_semiarid(), read_state(), read_urban_area(), read_urban_concentrations(), read_weighting_area()

Examples

# Read a particular grid at a given year
grid <- read_statistical_grid(code_grid = 45, year=2010)

# Read the grid covering a given state at a given year
state_grid <- read_statistical_grid(code_grid = "RJ")

[Package geobr version 1.9.1 Index]