read_weighting_area {geobr}	R Documentation

Download spatial data of Census Weighting Areas (area de ponderacao) of the Brazilian Population Census

Description

Only 2010 data is currently available.

Usage

read_weighting_area(
  code_weighting = "all",
  year = 2010,
  simplified = TRUE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

code_weighting
The 7-digit code of a Municipality. If the two-digit code or a two-letter uppercase abbreviation of a state is passed, (e.g. 33 or "RJ") the function will load all weighting areas of that state. If code_weighting="all", all weighting areas of the country are loaded.

year
Numeric. Year of the data. Defaults to 2010.

simplified
Logic FALSE or TRUE, indicating whether the function should return the data set with 'original' spatial resolution or a data set with 'simplified' geometry. Defaults to TRUE. For spatial analysis and statistics users should set simplified = FALSE. Borders have been simplified by removing vertices of borders using st_simplify{sf} preserving topology with a dTolerance of 100.

showProgress
Logical. Defaults to TRUE display progress bar.

cache
Logical. Whether the function should read the data cached locally, which is faster. Defaults to cache = TRUE. By default, geobr stores data files in a temporary directory that exists only within each R session. If cache = FALSE, the function will download the data again and overwrite the local file.

Value

An "sf" "data.frame" object

See Also

Other area functions: read_amazon(), read_biomes(), read_capitals(), read_comparable_areas(), read_country(), read_disaster_risk_area(), read_health_facilities(), read_health_region(), read_immediate_region(), read_indigenous_land(), read_intermediate_region(), read_meso_region(), read_metro_area(), read_micro_region(), read_municipal_seat(), read_municipality(), read_neighborhood(), read_pop_arrangements(), read_region(), read_schools(), read_semiarid(), read_state(), read_statistical_grid(), read_urban_area(), read_urban_concentrations()

Examples

# Read specific weighting area at a given year
w <- read_weighting_area(code_weighting=5201108005004, year=2010)

# Read all weighting areas of a state at a given year
w <- read_weighting_area(code_weighting=53, year=2010) # or
w <- read_weighting_area(code_weighting="DF", year=2010)
plot(w)

# Read all weighting areas of a municipality at a given year
w <- read_weighting_area(code_weighting=5201108, year=2010)
plot(w)

# Read all weighting areas of the country at a given year
w <- read_weighting_area(code_weighting="all", year=2010)

[Package geobr version 1.9.1 Index]